configured CAN bus to use mailboxes.
