+++ lib/util/heartbeat.cpp
@@ -1,21 +1,24 @@
- /// 
- ///  @ Author: Kevin Gilliam
- ///  @ Create Time: 2022-09-06 12:05:44
- ///  @ Modified by: Kevin Gilliam
- ///  @ Modified time: 2022-09-08 08:13:58
- ///  @ Description:
- ///
+///
+///  @ Author: Kevin Gilliam
+///  @ Create Time: 2022-09-06 12:05:44
+///  @ Modified by: Kevin Gilliam
+///  @ Modified time: 2022-09-08 12:40:48
+///  @ Description:
+///
 
 #include "heartbeat.h"
 #include <cctype>
 #include <cstdint>
 #include <device.h>
 
+#define HEARTBEATMODE_USE_TIMER 1
 enum ledState
 {
     LED_OFF = LOW,
     LED_ON = HIGH
 };
+
+IntervalTimer heartbeatTimer;
 
 static uint32_t cnt = 0;
 static ledState heartbeatState = LED_OFF;
@@ -23,6 +26,7 @@ uint32_t heartbeatPeriodCnts = 1000;
 
 void resetHeartbeat();
 ledState toggleHeartbeatState(ledState hbState);
+void toggleHeartbeatState();
 void setHeartbeatStateOff();
 void setHeartbeatStateOn();
 void updateLedPin();
@@ -33,7 +37,7 @@ void updateLedPin();
  */
 void initHeartbeat()
 {
-    pinMode(LED_PIN, OUTPUT);
+
     setHeartbeatStateOn();
     updateLedPin();
 }
@@ -57,6 +61,7 @@ void resetHeartbeat()
  */
 bool pingHeartBeat()
 {
+    #if !HEARTBEATMODE_USE_TIMER
     bool pinToggledFlag = false;
     if (cnt++ >= heartbeatPeriodCnts)
     {
@@ -66,11 +71,22 @@ bool pingHeartBeat()
     }
     updateLedPin();
     return (pinToggledFlag);
+    #else
+    return false;
+    #endif
 }
 
-void setHeartBeatPeriod(uint32_t cnts)
+void setHeartBeatPeriod(uint32_t timerPrd)
 {
+#if HEARTBEATMODE_USE_TIMER
+    heartbeatTimer.begin(toggleHeartbeatState, timerPrd);
+#else
     heartbeatPeriodCnts = cnts;
+#endif
+}
+void toggleHeartbeatState()
+{
+    TOGGLE_LED_PIN();
 }
 /**
  * @brief
