+++ platformio.ini
@@ -12,22 +12,29 @@
 platform = teensy
 board = teensy41
 framework = arduino
-; lib_dir = 
-; 	can_comms
-; build_flags = 
-; 	-I exec/
-; 	-I util/
+
+; [env:teensy41_A]
+; platform = teensy
+; board = teensy41
+; framework = arduino
+; upload_port = COM5
+
+; [env:teensy41_B]
+; platform = teensy
+; board = teensy41
+; framework = arduino
+; upload_port = COM6
 
 ; [env:native]
 ; platform = native
 ; test_ignore = test_embedded
 ; build_flags = -std=gnu++11
 
-[env:unitTestBuild]
-platform = teensy
-board = teensy41
-framework = arduino
-build_flags = 
-    -I .
-    -I ./test/
-lib_deps = throwtheswitch/Unity@^2.5.2
+; [env:unitTestBuild]
+; platform = teensy
+; board = teensy41
+; framework = arduino
+; build_flags = 
+;     -I .
+;     -I ./test/
+; lib_deps = throwtheswitch/Unity@^2.5.2
