--- src/exec/SetupExec.h
@@ -1,12 +1,0 @@
- /// 
- ///  @ Author: Kevin Gilliam
- ///  @ Create Time: 2022-09-06 10:42:24
- ///  @ Modified by: Kevin Gilliam
- ///  @ Modified time: 2022-09-07 16:40:55
- ///  @ Description:
- ///
-
-#include <Arduino.h>
-#include <FlexCAN_T4.h>
-
-void setupExec(void);
